tests/wpt/web-platform-tests/IndexedDB/idbobjectstore_get6.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBObjectStore.get() - throw TransactionInactiveError on aborted transaction </title>
<link rel="author" title="YuichiNukiyama" href="https://github.com/YuichiNukiyama">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>
    var db,
      t = async_test();

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        db.createObjectStore("store", { keyPath: "key" })
    }

    open_rq.onsuccess = function (e) {
        var store = db.transaction("store")
                      .objectStore("store");
        store.transaction.abort();
        assert_throws("TransactionInactiveError", function () {
            store.get(1);
        }, "throw TransactionInactiveError on aborted transaction.");
        t.done();
    }
</script>

<div id="log"></div>
